Revert "feature(): add project pages #52"
