+++ package.json
@@ -12,31 +12,26 @@
     "test": "jest"
   },
   "dependencies": {
-    "@fec/remark-a11y-emoji": "^4.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-markdown": "^6.0.3",
-    "react-router-dom": "^6.21.1",
+    "react-router-dom": "^6.13.0",
     "react-simple-typewriter": "^5.0.1",
-    "rehype-highlight": "^4.0.0",
-    "remark-gfm": "^1.0.0",
-    "remark-toc": "^7.0.0",
-    "styled-components": "^6.1.6"
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.1.6",
-    "@testing-library/react": "^14.1.2",
-    "@types/jest": "^29.5.11",
-    "@types/react": "^18.2.46",
-    "@types/react-dom": "^18.2.18",
-    "@types/styled-components": "^5.1.34",
-    "@vitejs/plugin-react": "^4.2.1",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-styled-components": "^7.2.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.3.3",
-    "vite": "^5.0.10"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.2.5",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@types/styled-components": "^5.1.26",
+    "@vitejs/plugin-react": "^3.0.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-styled-components": "^7.1.1",
+    "ts-jest": "^29.0.5",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.0"
   }
 }
