+++ src/assets/data/projectData.json
@@ -1,31 +1,14 @@
 {
   "projectData": [
     {
-      "id": 15,
+      "id": 14,
       "title": "About Me API",
       "details": "Developed a API so others can learn about who I am and what I'm currently doing using Rust, Rocket, and MySQL",
       "skills": ["Rust", "Rocket.rs", "MySQL", "Shuttle.rs", "SQL"],
       "topics": ["software_engineering"],
       "isRepo": true,
       "isWebApp": false,
-      "link": ""
-    },
-    {
-      "id": 14,
-      "title": "Wikipedia Information Retrieval System",
-      "details": "Developed a information retrieval system that uses the vector space model to rank Wikipedia documents based on a query",
-      "skills": [
-        "Python",
-        "Numpy",
-        "Pandas",
-        "NLTK",
-        "Scikit-Learn",
-        "Streamlit"
-      ],
-      "topics": ["data_science"],
-      "isRepo": false,
-      "isWebApp": true,
-      "link": ""
+      "link": "#"
     },
     {
       "id": 13,
@@ -38,8 +21,11 @@
         "Node",
         "TypeScript",
         "Styled-Components",
+        "React Router",
+        "Recharts",
         "SQL",
-        "Github Actions"
+        "Github Actions",
+        "Docker"
       ],
       "topics": ["software_engineering", "web_development"],
       "isRepo": true,
@@ -49,7 +35,7 @@
     {
       "id": 12,
       "title": "Portfolio Website",
-      "details": "Created a website to highlight projects and person brand with React, TypeScript, and Styled-Components",
+      "details": "Created a web application/page to highlight projects and person brand with React, TypeScript, and Styled-Components",
       "skills": [
         "React",
         "TypeScript",
@@ -64,7 +50,7 @@
     },
     {
       "id": 11,
-      "title": "Finding Features About New Galaxies using ML",
+      "title": "Finding Features About New Galaxies with Machine Learning",
       "details": "Trained and tested 4 supervised machine learning models to predict 4 features of galaxy from a spectra of the same galaxy: the total amount of mass in stars with a galaxy, the amount of cosmic dust in a galaxy, how long ago did the galaxy form its first stars, and the speed of the process of star formation",
       "skills": ["Python", "Numpy", "Matplotlib", "Seaborn", "Scikit-Learn"],
       "topics": ["machine_learning", "data_science"],
