--- src/assets/markdown/projects/project_1.md
@@ -1,37 +1,0 @@
-# Life Expectancy and GDP Analysis
-
-Analyzed a dataset containing life expectancy and GDP of 6 countries across the world with different circumstances
-
-## Table of Contents
-
-## Introduction
-
-Five questions were asked and answered in this project:
-
-1. Is there disparity between each country with GDP and Life Expectancy?
-
-2. Is there correlation between GDP and life expectancy for each country?
-
-3. Has GDP increased for each country?
-
-4. Has life expectancy risen for each country?
-
-5. What's the average life expectancy between these countries?
-
-## Results
-
-1. There disparity between each country for both GDP and Life Expectancy. The distribution of GDP among the six countries was skewed right, meaning within the sample the mean doesn't represent the population well. The distribution of life expectancy among the six countries was skewed to the left.
-
-2. There is positive correlation between GDP and life expectancy for every country in the data set.
-3. The GDP has increased for each country with different rates each country for each year Some countries shown higher rates of change of their GDP over the 15 year period. Two examples are USA and China
-
-4. The life expectancy has increased for each country with different rates each country for each year Some countries show higher rates of change of their life expectancy over the 15 year period. One example is Zimbabwe.
-
-5. The average life expectancy between the majority of countries in the sample is between 70 and 80 years while Zimbabwe being the outlier with average of 50 years.
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Seaborn
