--- src/assets/markdown/projects/project_10.md
@@ -1,30 +1,0 @@
-# NYC Airbnb Listing Price Predictor App
-
-A web application that allows users to find out how much a listing would cost based on the features of the listing.
-
-## Table of Contents
-
-## Data
-
-The data was obtained from [Kaggle](https://www.kaggle.com/dgomonov/new-york-city-airbnb-open-data). The data was cleaned and preprocessed using pandas and matplotlib.
-
-## Modeling
-
-A naive bayes classifier was used to classify the tweets after they were vectorized using TF-IDF vectorization
-
-## GitHub Repo
-
-[GitHub Repo](https://github.com/112523chen/CET4973-Fall2022-Project)
-
-## Demo
-
-[Live Demo](https://112523chen-cet4973-fall2022-project-main-o7ewsh.streamlit.app/)
-
-## Technologies Used
-
-- Python
-- Numpy
-- Pandas
-- Matplotlib
-- Scikit-learn
-- Streamlit
