--- src/assets/markdown/projects/project_11.md
@@ -1,38 +1,0 @@
-# Finding Features About New Galaxies using Machine Learning
-
-A project that uses machine learning to find features about new galaxies just using their spectra.
-
-## Table of Contents
-
-## Data
-
-The data was obtained from a professor at New York City College of Technology Physics Department.
-
-- Targets
-
-  - Age: how long ago did the galaxy form its first stars?
-
-  - Tau: how fast/slow the process of star formation proceeds.
-
-  - Stellar Mass: the total amount of mass in stars in the galaxy.
-
-  - Dust (A_v): Amount of cosmic dust in a galaxy, measured as the diminishing of optical light.
-
-- Features
-
-  - Spectra: the light from a galaxy broken down into its component wavelengths.
-
-## Modeling
-
-The data was split into training and testing sets. The training set was used to train the model and the testing set was used to test the model. The model was trained using the training set and then tested using the testing set. The model was then evaluated using the mean squared error (MSE) and the R^2 score. The model was then used to predict the targets of the testing set.
-
-## GitHub Repo
-
-[GitHub Repo](https://github.com/112523chen/ML-Galaxy-Project/blob/master/Research.ipynb)
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Scikit-learn
