--- src/assets/markdown/projects/project_12.md
@@ -1,71 +1,0 @@
-# Personal Website
-
-A personal website that showcases my skills, projects, and learning experiences.
-
-## Table of Contents
-
-## Technologies Used Through Life of Project
-
-- React
-- JavaScript
-- TypeScript
-- HTML
-- CSS
-- SASS
-- Styled Components
-- GitHub Pages
-- GitHub Actions
-
-## Demos
-
-- [You are already here 😄](https://112523chen.github.io)
-
-- [Version 3.1 Live Demo](https://112523chen.github.io/PortfolioWebsite3.1/)
-
-- [Version 3.0 Live Demo](https://112523chen.github.io/PortfolioWebsite3.0/)
-
-- [Version 2.5 Live Demo](https://112523chen.github.io/PortfolioWebsite2.5/)
-
-- [Version 1 Live Demo](https://112523chen.github.io/PortfolioWebsite1.0/)
-
-## Change Log
-
-### [Current Version](https://github.com/112523chen/112523chen.github.io)
-
-- [x] Add contact section to the website.
-- [x] Add modular markdown pages for each project to be rendered into the pages.
-- [x] Update side bar CSS to be more esthetically pleasing.
-- [x] Update project section with new pages.
-- [x] Refactor project's directory structure.
-- [x] Refactor components to be more modular.
-
-### [Version 3.3](https://github.com/112523chen/112523chen.github.io/releases/tag/v3.3)
-
-- [x] Add work experience section to the website.
-- [x] Add what I'm doing now section to the website.
-- [x] Add side bar to the website for navigation.
-
-### [Version 3.2](https://github.com/112523chen/112523chen.github.io/releases/tag/v3.2)
-
-- [x] Refactor project directory structure.
-
-### [Version 3.1](https://github.com/112523chen/PortfolioWebsite3.1)
-
-- [x] Convert CSS into React Styled Components.
-
-### [Version 3.0](https://github.com/112523chen/PortfolioWebsite3.0)
-
-- [x] Redesign the website to be more modern and professional.
-- [x] Convert from vanilla JavaScript to React.
-- [x] Switch from CSS to SCSS.
-- [x] Switch from JavaScript to TypeScript.
-- [x] Add tests to the website.
-- [x] Add GitHub Actions to the website.
-
-### [Version 2.5](https://github.com/112523chen/PortfolioWebsite2.5)
-
-- [x] Add a new project section to the website.
-
-### [Version 1](https://github.com/112523chen/PortfolioWebsite1.0)
-
-- [x] Create a personal website that showcases my skills, projects, and learning experiences.
