--- src/assets/markdown/projects/project_13.md
@@ -1,29 +1,0 @@
-# Job Tracker Web App
-
-An interactive web app that visualizes user job applications data using React and Express.
-
-## Table of Contents
-
-## Introduction
-
-An interactive web app that allows the user to keep track of their job applications. The app allows the user to add, edit, and delete job applications. The app also allows the user to filter their job applications by company name, job title, and job status.
-
-## Installation
-
-```bash
-git clone https://github.com/112523chen/RedditClientWebApp.git
-```
-
-## Repo
-
-[Job Tracker App](https://github.com/112523chen/Job-Tracker-Web-App)
-
-## Technologies Used
-
-- React
-- TypeScript
-- Node.js
-- Express
-- Postgres
-- SQL
-- Github Actions
