--- src/assets/markdown/projects/project_14.md
@@ -1,31 +1,0 @@
-# Wikipedia Information Retrieval System
-
-A web application that allows users to search for Wikipedia articles and retrieve relevant information.
-
-## Table of Contents
-
-## Introduction
-
-## Installation
-
-```bash
-git clone https://github.com/112523chen/wikipedia-search-engine-web-app.git
-```
-
-## Demo
-
-TODO: Add demo
-
-## Repo
-
-[Wikipedia Information Retrieval System](https://github.com/112523chen/wikipedia-search-engine-web-app)
-
-## Technologies Used
-
-- Python
-- Streamlit
-- Scikit-learn
-- NLTK
-- Wikipedia API
-- Ollama API
-- llama2
