--- src/assets/markdown/projects/project_15.md
@@ -1,23 +1,0 @@
-# About ME API
-
-A RESTful API that returns information about me.
-
-## Table of Contents
-
-## Introduction
-
-This API is built using Rocket.rs and is hosted on [Shuttle.rs](https://www.shuttle.rs/). It returns information about me, including my name, age, and interests.
-
-## Routes
-
-## Repo
-
-[About ME API](#)
-
-## Technologies Used
-
-- Rust
-- Rocket.rs
-- Shuttle.rs
-- MySQL
-- Planet Scale
