--- src/assets/markdown/projects/project_2.md
@@ -1,26 +1,0 @@
-# Restaurant Website
-
-A website that allows users to look through a menu of food items and location for a restaurant.
-
-## Table of Contents
-
-## Introduction
-
-A website that is hosted on GitHub[^1] that allows users to look through a menu of food items and location for a made up restaurant.
-
-## Installation
-
-```bash
-git clone https://github.com/112523chen/RestaurantWebsite.git
-```
-
-## Usage
-
-[Live Demo](https://112523chen.github.io/RestaurantWebsite/)
-
-## Technologies Used
-
-- HTML
-- CSS
-
-[^1]: [GitHub Repo](https://github.com/112523chen/RestaurantWebsite)
