--- src/assets/markdown/projects/project_3.md
@@ -1,27 +1,0 @@
-# Gym Website
-
-A website that is hosted on GitHub[^1] that allows users to look at the different services offered by a made up gym.
-
-## Table of Contents
-
-## Introduction
-
-An interactive website for a gym brand to illustrate and deepen HTML, CSS, and JavaScript skills, along with the concept of box models within CSS
-
-## Installation
-
-```bash
-git clone https://github.com/112523chen/GymWebsite.git
-```
-
-## Usage
-
-[Live Demo](https://112523chen.github.io/GymWebsite/)
-
-## Technologies Used
-
-- HTML
-- CSS
-- JavaScript
-
-[^1]: [GitHub Repo](https://github.com/112523chen/GymWebsite)
