--- src/assets/markdown/projects/project_4.md
@@ -1,20 +1,0 @@
-# Manhattan School Districts Standardized Testing Analysis
-
-Create a visual to display the math and ELA scores of standardized testing throughout Manhatten to help highlight areas of Manhatten where standardized needs to improve.
-
-## Table of Contents
-
-## Open Questions
-
-1. Why is the test score for both the math and ELA state tests in 2018 so high compared to past years?
-
-2. Why are schools having a mean math and ELA score lower than the average for their district?
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Seaborn
-
-[GitHub Repo](https://github.com/112523chen/School-Districts-Standardized-Test-Analysis)
