--- src/assets/markdown/projects/project_5.md
@@ -1,50 +1,0 @@
-# Biodiversity in National Parks Analysis
-
-Analyzed a dataset containing life expectancy and GDP of 6 countries across the world with different circumstances
-
-## Table of Contents
-
-## Introduction
-
-The goal of this project is to analyze biodiversity data from the National Parks Service, with special emphasis on the various species observed in the different national Park locations.
-
-This project will scope, analyze, prepare, plot data, and seek to explain the findings from the analysis
-
-Here are a few questions that this project intends to answer
-
-What is the distribution of conservation status for species?
-Are certain species more likely to be endangered?
-Which animal is most common and what is their distribution amongst the parks?
-Are the difference between species and their conservation status significant?
-Data Sources: Both `Observations.csv` and `Species_info.csv` was provided by Codecademy
-
-Note: The data within this project is inspired by real data and doesn't 100 percent accurately describe the real data in the real world
-
-## Results
-
-Within this project, several data visualizations and inferences about the various in four of the National Parks presented in this data set
-
-Some of the questions that were able to be answered now:
-
-What is the distribution of conservation status for species?
-
-> The major of species without data set where not part of conservation: 191 were and 5,633 weren't.
-
-Are certain species more likely to be endangered?
-
-> Birds and Mammals had the highest percentage of being in protection.
-
-Which animal is most common and what is their distribution amongst the parks?
-
-> Through analysis, bats were founded to be that most common animal which the majority of bats in our data set residing in Yellowstone National Park.
-
-Are the difference between species and their conservation status significant?
-
-> There are some cases such as for mammals and reptiles and Reptile Nonvascular Plant
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Seaborn
