--- src/assets/markdown/projects/project_6.md
@@ -1,26 +1,0 @@
-# Reddit Client App
-
-An interactive web app that allows users to look at the top 50 posts on Reddit.
-
-## Table of Contents
-
-## Introduction
-
-An interactive web app that allows users to look at the top 50 posts on Reddit. The app uses the Reddit JSON API to fetch the latest posts from several different subreddits.
-
-## Installation
-
-```bash
-git clone https://github.com/112523chen/RedditClientWebApp.git
-```
-
-## Usage
-
-[Live Demo](https://112523chen.github.io/RedditClientWebApp/)
-
-## Technologies Used
-
-- React
-- HTML
-- CSS
-- JavaScript
