--- src/assets/markdown/projects/project_7.md
@@ -1,31 +1,0 @@
-# Image Classification App
-
-A web application that allows users to see what natural scenes are in a given image.
-
-## Table of Contents
-
-## Data
-
-[Intel Image Classification](https://www.kaggle.com/datasets/puneet6060/intel-image-classification)
-
-> A dataset from Kaggle that contains over 17,000 images of buildings, forests, glaciers, mountains, bodies of water, and streets
-
-## Modeling
-
-A random forest classifier was used to classify the images.
-
-## GitHub Repo
-
-[GitHub Repo](https://github.com/112523chen/Image-Classification-App)
-
-## Demo
-
-[Live Demo](https://112523chen-image-classification-dem0.streamlit.app/)
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Scikit-learn
-- Streamlit
