--- src/assets/markdown/projects/project_8.md
@@ -1,28 +1,0 @@
-# Emotion Detection App
-
-A web application that allows users to see what emotion is in a given tweet.
-
-## Table of Contents
-
-## Data
-
-## Modeling
-
-A naive bayes classifier was used to classify the tweets after they were vectorized using TF-IDF vectorization
-
-## GitHub Repo
-
-[GitHub Repo](https://github.com/112523chen/Project_SA_CTP)
-
-## Demo
-
-[Live Demo](https://112523chen-text-emotion-classification-ctp-dem0.streamlit.app/)
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Scikit-learn
-- Streamlit
-- NLTK
