--- src/assets/markdown/projects/project_9.md
@@ -1,20 +1,0 @@
-# NYC Airbnb Listings Analysis
-
-A project that analyzes NYC Airbnb listings data to find meaningful insights.
-
-## Table of Contents
-
-## Data
-
-The data was obtained from [Kaggle](https://www.kaggle.com/dgomonov/new-york-city-airbnb-open-data). The data was cleaned and preprocessed using pandas and matplotlib.
-
-## GitHub Repo
-
-[GitHub Repo](https://github.com/112523chen/NYC_Airbnb_Analysis)
-
-## Technologies Used
-
-- Python
-- Pandas
-- Matplotlib
-- Seaborn
