+++ src/components/DisplayButton/DisplayButton.style.tsx
@@ -7,18 +7,13 @@ interface ButtonProps {
   hoverColor: string;
 }
 
-interface ModeDisplayProps {
-  useInMainPage?: boolean;
-}
-
-export const ModeDisplay = styled.div<ModeDisplayProps>`
-  ${(props) => (props.useInMainPage ? "position: absolute;" : "")}
-  ${(props) => (props.useInMainPage ? "top: 6%;" : "")}
-  ${(props) => (props.useInMainPage ? "right: 6%;" : "")} 
-  ${(props) =>
-    props.useInMainPage
-      ? "@media only screen and (max-width: 24rem) {right: 10%;}"
-      : ""}
+export const ModeDisplay = styled.div`
+  position: absolute;
+  right: 6%;
+  top: 6%;
+  @media only screen and (max-width: 24rem) {
+    right: 10%;
+  }
 `;
 
 export const Button = styled.button<ButtonProps>`
