+++ src/components/ProjectsAlt/ProjectsAlt.tsx
@@ -4,11 +4,14 @@ import {
   ProjectAltBase,
   ProjectAltDetails,
   ProjectAltLI,
+  ProjectAltLink,
+  ProjectAltLinkButton,
   ProjectAltSkills,
   ProjectAltText,
   ProjectAltTitle,
   ProjectAltUL,
 } from "./ProjectsAlt.style";
+import { getLinkText, getLinkTitle } from "../../functions/helper";
 
 interface Props {
   project: ProjectType;
@@ -17,14 +20,10 @@ interface Props {
 
 const ProjectsAlt: React.FC<Props> = ({ project, mode }) => {
   return (
-    <ProjectAltBase mode={mode} to={`/project/${project.id}`}>
+    <ProjectAltBase mode={mode}>
       <ProjectAltTitle>{project.title}</ProjectAltTitle>
       <ProjectAltDetails>
-        <ProjectAltText>
-          {project.details.length > 200
-            ? project.details.substring(0, 200) + "..."
-            : project.details}
-        </ProjectAltText>
+        <ProjectAltText>{project.details}</ProjectAltText>
       </ProjectAltDetails>
       <ProjectAltSkills>
         <ProjectAltUL>
@@ -35,6 +34,16 @@ const ProjectsAlt: React.FC<Props> = ({ project, mode }) => {
           ))}
         </ProjectAltUL>
       </ProjectAltSkills>
+      <ProjectAltLinkButton mode={mode}>
+        <ProjectAltLink
+          mode={mode}
+          title={getLinkTitle(project.isRepo, project.isWebApp)}
+          target="_blank"
+          href={project.link}
+        >
+          {getLinkText(project.isRepo, project.isWebApp)}
+        </ProjectAltLink>
+      </ProjectAltLinkButton>
     </ProjectAltBase>
   );
 };
