+++ src/components/model.ts
@@ -8,7 +8,6 @@ export interface ProjectType {
   isWebApp: boolean;
   link: string;
   topics: string[];
-  markdown?: string;
 }
 
 //* Definition of a theme state
