+++ src/pages/ProjectPage/ProjectPage.tsx
@@ -1,18 +1,16 @@
 import React, { useEffect } from "react";
-import { LightState, ProjectType } from "../../components/model";
-import { useParams } from "react-router-dom";
+import { LightState, ProjectType, TopicType } from "../../components/model";
+import ProjectAlt from "../../components/ProjectsAlt/ProjectsAlt";
 import {
+  FilterLink,
+  FrameAlt,
   HomeLink,
-  Content,
-  ProjectPageBase,
-  NavBar,
-  ContentWrapper,
+  Title,
+  TopicFrame,
 } from "./ProjectPage.style";
-import Markdown from "react-markdown";
-import remarkGfm from "remark-gfm";
-import rehypeHighlight from "rehype-highlight";
-import remarkToc from "remark-toc";
 import DisplayButton from "../../components/DisplayButton/DisplayButton";
+import { useLocation, useParams } from "react-router-dom";
+import { getFilterLabel } from "../../functions/helper";
 
 interface Props {
   projectData: ProjectType[];
@@ -21,58 +19,60 @@ interface Props {
 }
 
 const ProjectPage: React.FC<Props> = ({ projectData, mode, setMode }) => {
-  const params = useParams() as { id: string };
-  const [markdownFile, setMarkdownFile] = React.useState<string>("");
+  const [projects, setProjects] = React.useState<ProjectType[]>(projectData);
 
-  const project: ProjectType = projectData.filter(
-    (project) => project.id === Number(params.id)
-  )[0];
+  const params = useParams();
+  const location = useLocation();
 
   useEffect(() => {
-    import(`../../assets/markdown/projects/project_${params.id}.md`)
-      .then((module) => {
-        setMarkdownFile(module.default);
-      })
-      .catch((error) => {
-        console.log(`Failed to load markdown file:\n${error}`);
-      });
-  });
+    if (params.hasOwnProperty("type")) {
+      setProjects(
+        projectData.filter((project) =>
+          project.topics.includes(params.type as string)
+        )
+      );
+    } else {
+      setProjects(projectData);
+    }
 
-  if (!project) {
-    return (
-      <ProjectPageBase mode={mode}>
-        <NavBar mode={mode}>
-          <HomeLink mode={mode} className="link" to="/projects">
-            Go Back
-          </HomeLink>
-          <DisplayButton mode={mode} setMode={setMode} useInMainPage={false} />
-        </NavBar>
-        <div style={{ textAlign: "center" }}>
-          <h1>404: Page Not Found</h1>
-        </div>
-      </ProjectPageBase>
-    );
-  }
+    window.scrollTo(0, 0);
+  }, [location.pathname]);
 
   return (
-    <ProjectPageBase mode={mode}>
-      <NavBar mode={mode}>
-        <HomeLink mode={mode} className="link" to="/projects">
-          Go Back
-        </HomeLink>
-        <DisplayButton mode={mode} setMode={setMode} useInMainPage={false} />
-      </NavBar>
-      <ContentWrapper mode={mode}>
-        <Content>
-          <Markdown
-            remarkPlugins={[remarkGfm, remarkToc]}
-            rehypePlugins={[rehypeHighlight]}
-          >
-            {markdownFile}
-          </Markdown>
-        </Content>
-      </ContentWrapper>
-    </ProjectPageBase>
+    <div
+      style={{
+        backgroundColor: mode === "light" ? "#fcfcfc" : "#111827",
+        minHeight: "100vh",
+      }}
+    >
+      <HomeLink mode={mode} className="link" to="/">
+        Go Back
+      </HomeLink>
+      <DisplayButton mode={mode} setMode={setMode} />
+      <Title id="title">Project Page</Title>
+      <TopicFrame>
+        {Object.values(TopicType)
+          .filter((v) => isNaN(Number(v)))
+          .map((topic) => (
+            <FilterLink
+              mode={mode}
+              key={topic}
+              to={`/projects/${topic !== "all" ? topic : ""}`}
+            >
+              {getFilterLabel(topic as string)}
+            </FilterLink>
+          ))}
+      </TopicFrame>
+      <FrameAlt>
+        {projects.length > 0 ? (
+          projects.map((project) => (
+            <ProjectAlt project={project} key={project.id} mode={mode} />
+          ))
+        ) : (
+          <h1>No projects found</h1>
+        )}
+      </FrameAlt>
+    </div>
   );
 };
