+++ src/vite-env.d.ts
@@ -1,2 +1,1 @@
 /// <reference types="vite/client" />
-declare module "*.md";
