+++ vite.config.ts
@@ -4,16 +4,5 @@ import react from "@vitejs/plugin-react";
 // https://vitejs.dev/config/
 export default defineConfig({
   base: "/",
-  plugins: [
-    react(),
-    {
-      name: "markdown-loader",
-      transform(code, id) {
-        if (id.slice(-3) === ".md") {
-          // For .md files, get the raw content
-          return `export default ${JSON.stringify(code)};`;
-        }
-      },
-    },
-  ],
+  plugins: [react()],
 });
